# docs/openapi.yaml
openapi: 3.0.3
info:
  title: spin.clinic API
  version: 1.0.0
  description: 'REST API spin.clinic: pasek materiałów i konfiguracja portalu, materiały
    i ich kontekst, nitki, reakcje, rejestr osób publicznych oraz narzędzia redakcji.
    Każdy materiał ma źródło, datę i odnośnik do oryginału.'
paths:
...
  /api/staff/political/posts/:
    get:
      operationId: staff_political_posts_list
      description: Read-only review surface for the editorial panel. Serves only already-stored
        posts.
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - staff
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPoliticalPostList'
          description: ''
...
  /api/staff/political/status/:
    get:
      operationId: staff_political_status_retrieve
      summary: Stan pobierania kont politycznych (redakcja)
      tags:
      - redakcja
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
    AccountCampEnum:
      enum:
      - government
      - opposition
      - public
      type: string
      description: |-
        * `government` - Obóz rządzący
        * `opposition` - Opozycja
        * `public` - Instytucja publiczna
...
    PaginatedPoliticalPostList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PoliticalPost'
...
    PoliticalPost:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          type: integer
          readOnly: true
        account_handle:
          type: string
          readOnly: true
        account_display_name:
          type: string
          readOnly: true
        post_id:
          type: string
          readOnly: true
        url:
          type: string
          format: uri
          readOnly: true
        text:
          type: string
          readOnly: true
        published_at:
          type: string
          format: date-time
          readOnly: true
        fetched_at:
          type: string
          format: date-time
          readOnly: true
        author_data:
          readOnly: true
        media:
          readOnly: true
        camp_at_collection:
          allOf:
          - $ref: '#/components/schemas/AccountCampEnum'
          readOnly: true
        available:
          type: boolean
          readOnly: true
      required:
      - account
      - account_display_name
      - account_handle
      - author_data
      - available
      - camp_at_collection
      - fetched_at
      - id
      - media
      - post_id
      - published_at
      - text
      - url
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid